Redid Menu System
Created the new menu system for the overhaul. Can now transition from scenes without crashing, and is ready for the vehicle and stage menus to be implemented when we add those in. If not then it is working as it is, just need to fix the GameManager to initialize variables for levels