Touch ups to GameManager, UIManager, and Menus
Made Pause, Instructions, and Settings to have a dark but transpearent background to show that the game is paused. The GameManager and PlayerControllers now get paused appropriately, and players cannot move when game is paused. The game timer will also be paused, but there is a glitch that needs to be fixed, the timer GameObject cannot be found by the GameManager. Possible fix could be to just have a LevelManager. Other than that, this build is working.